backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Use SQLite instead of PostgreSQL for easier setup
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9
# alembic==1.13.0

# AI/ML Libraries (install main ones first)
openai==1.3.0
groq==0.31.1
anthropic==0.34.0
transformers==4.44.0
torch==2.0.1
numpy==1.24.3
scikit-learn==1.3.0

# Vector Databases (simplified)
chromadb==0.4.18

# Search and Retrieval
redis==5.0.1

# Web Scraping and Processing
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1

# Privacy and Encryption
cryptography==41.0.7

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
pillow==10.1.0

# Enhanced RAG System Dependencies
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Data Processing
pandas==2.1.3

# Environment and Configuration
python-dotenv==1.0.0

# RSS and Feed Processing
feedparser==6.0.10

# Development and Testing
pytest==7.4.3
httpx==0.25.2